SanMartino Campanaro Changes
AudioRecorder: added a lastStartTime variable to keep track of consecutive calls to the split method, to report Timeout even within multiple calls.
Added a method to reset the timeout to be called after the robot talked to reset the window for the timeout

DialogueState: added a method to reset the conversation history, to be called when a new interaction sequence is started

MainActivity: introduced an onGoingIntervention variable to keep track of a DueIntervention that spans multiple phrases (interaction_sequence).
Introduced the interaction sequence intervention.
Also, now we send a modified sentence in case of timeout to be recognized server side, if the intervention is of type interaction sequence.
At the end of a sequence of reply/continuation, the audioRecord timeout is reset.
Furthermore, the handle() now advances the intervention if of type interaction sequence.

PersonalizationServer: added the interaction sequence intervention type.

ServerCommunicatioManager: added support for the contextual_data needed by the intervention interaction sequence.